move tests to netcoreapp specific file
